Refactor code with shaper in a new file
