+++ constants.js
@@ -1,0 +1,7 @@
+export const colors = ['#1abc9c', '#2ecc71', '#3498db', '#9b59b6', '#34495e', '#16a085', '#27ae60', '#2980b9', '#8e44ad', '#2c3e50',
+    '#f1c40f', '#e67e22', '#e74c3c', '#ecf0f1', '#95a5a6', '#f39c12', '#d35400', '#c0392b', '#bdc3c7', '#7f8c8d'];
+
+export const margin = { top: 15, right: 10, bottom: 15, left: 10 };
+export const width = 1350 - margin.left - margin.right;
+export const height = 750 - margin.top - margin.bottom;
+export const strokeWidth = 0.5;
